requirements.txt
requests
pandas
websockets
psycopg2-binary
prefect
